Moved painting series point markers logic to a separate module
